domains/Get_News/pyproject.toml
[tool.poetry]
name = "news-summarization"
version = "0.1.0"
description = ""
authors = ["Tan Chau <thomas.lezard.2002@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "3.10.9"
fastapi = "0.100.0"
uvicorn = "0.22.0"
kafka-python = "2.0.2"
selenium = "4.10.0"
confluent-kafka = "2.2.0"
newspaper3k = "0.2.8"
mysql-connector-python = "8.0.33"
hdfs = "2.7.0"
aiohttp = "3.8.4"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
